streamlining + camelCasing
Removed every time (Except for constructors) where a Method or Property started with it's respective Class name (hopefully).

Made it so that with the exception of abbreviations (such as PP) all Methods and Properties use camelCase (this was mostly already done but removing the starting word kind of undid it).