Added itemAll() to Display
Added the ability to display all the information on an Item from the "Item" Abstract Class.

This notably does *not* display all the information provided by the Derived Class. I intend to implement this in the future but for now it's probably not worth the effort while I have other stuff to work on.