Battle now takes Trainers (Kind of)
Alright so now that we have implemented Trainers which can have Pokemon it makes more sense for Battle to use Trainers with Pokemon than just Pokemon. With this in mind Battle now uses Trainers and will go until one of them runs out of Pokemon.

There is currently a problem where instead of exiting normally after the last Pokemon faints an error is thrown (from needing to access an empty map).

A lot of this code can probably be extracted into much cleaner Functions/Methods but for now it works.

Definitely need to rework the whole way that individual Pokemon are getting referenced because this 'fasterPokemon', 'slowerPokemon', 'fasterPokemonOwner', 'trainer1Pokemon' thing is a very bad way to go about this.