Reverted use of DateTime
Using DateTime apparently causes a slew of issues. Two of these issues that come to mind are 1) If the user changes their device time while the program is running it can mess things up (minor issue, I know) 2) If the user pauses the game (not yet implemented) then the cooldown will reset despite no time having passed in game.

Because of all this I have decided to go back to the 'movementCooldown' method which is updated in 'FixedUpdate'.