Added Species!
Alright so this was a pain. Newtonsoft does seem to work fairly well but it was unintuitive to have to make nested classes to properly parse my data.

Now because I have Species I can implement much more that is dependent on it (like the "Evolve" Method in 'Pokemon' which needs to know what you're evolving into).