Now detects if position can spawn pokemon
After the player moves the new Bool 'playerMovedLastFrame' is set to True so that in the next iteration of 'FixedUpdate' it will log the movement. This has to be done (rather than checking if the player moved in the current iteration) because the player doesn't actually move until 'FixedUpdate' is called again.